添加了TestSeg类,这可以直接将c++代码复制到一个concrete类内后直接进行测试 添加了TestCasesManager管理一组solution指针，指向的不同对象分别进行运行测试 添加了Solution内一个成员函数string GetName()作为这个类测试名称的标记方法

+++ LeetCodeTest/TestSeg.h
@@ -75,7 +75,7 @@ public:
 	}
 	int RunAllSeg()
 	{
-		for (int i = 0; i < TestSegContainer.size(); ++i)
+		for (size_t i = 0; i < TestSegContainer.size(); ++i)
 		{
 			if (RunSegAt(i))
 			{
